fix: typo bugs with if/else statement - create directories respect the functionality - create state machine with the function is_opening_or_closing_quote() , update_quote_state()
