+++ Makefile
@@ -1,24 +1,36 @@
-CC = cc
-CFLAGS = -Wall -Wextra -Werror
-TARGET = minishell
-SRCS = main.c lexer.c lexer_utils.c parser.c parser_utils.c print_functions.c validate.c
-OBJS = $(SRCS:.c=.o)
+CC      = cc
+CFLAGS  = -Wall -Wextra -Werror -I include -I libft
+NAME    = minishell
 
-all: $(TARGET)
+SRCS    =   main.c \
+            print_functions.c \
+            src/lexer/lexer.c \
+            src/lexer/lexer_utils.c \
+            src/lexer/lexer_quotes.c \
+            src/parser/parser.c \
+            src/parser/parser_utils.c \
+            src/validate/validate.c
 
-$(TARGET): $(OBJS) libft/libft.a
-	$(CC) $(CFLAGS) -o $(TARGET) $(OBJS) libft/libft.a -lreadline
+OBJS    = $(SRCS:.c=.o)
+
+all: $(NAME)
+
+$(NAME): $(OBJS) libft/libft.a
+	$(CC) $(CFLAGS) -o $(NAME) $(OBJS) libft/libft.a -lreadline
 
 libft/libft.a:
-	make -C libft
+	$(MAKE) -C libft
 
 %.o: %.c
-	$(CC) $(CFLAGS) -c $< -o $@
+	$(CC) $(CFLAGS) -I include -c $< -o $@
 
-re: clean all
+re: fclean all
 
 clean:
-	rm -f $(OBJS) $(TARGET)
-	make -C libft clean
+	rm -f $(OBJS)
+	$(MAKE) -C libft clean
 
-.PHONY: all re clean
+fclean: clean
+	rm -f $(NAME)
+
+.PHONY: all re clean fclean
